+++ libs/gui/src/CDisplayWindowGUI.cpp
@@ -138,4 +138,153 @@ bool CDisplayWindowGUI::keyboardEvent(
 	return false;
 }
 
+nanogui::Window* CDisplayWindowGUI::createManagedSubWindow(
+	const std::string& title)
+{
+	constexpr std::size_t MAX_MINIMIZED_TITLE_LEN = 20;
+
+	// Create subwindow:
+	auto w = new nanogui::Window(this, title);
+	m_subWindows.windows.push_back(w);
+	const int thisWinIndex = m_subWindows.windows.size() - 1;
+
+	// Create UI on first call:
+	createSubWindowsControlUI();
+
+	// Append to combo:
+	ASSERT_(m_subWindows.uiCombo);
+	auto items = m_subWindows.uiCombo->items();
+	auto itemsShort = m_subWindows.uiCombo->itemsShort();
+
+	items.push_back(title);
+	itemsShort.push_back(title.substr(0, MAX_MINIMIZED_TITLE_LEN));
+
+	const auto oldSelected = m_subWindows.uiCombo->selectedIndex();
+
+	m_subWindows.uiCombo->setItems(items, itemsShort);
+	m_subWindows.uiCombo->setSelectedIndex(oldSelected > 0 ? oldSelected : 0);
+
+	// Add minimize/maximize buttons:
+	w->buttonPanel()
+		->add<nanogui::Button>("", ENTYPO_ICON_MINUS)
+		->setCallback(
+			[this, thisWinIndex]() { m_subWindows.minimize(thisWinIndex); });
+
+	return w;
+}
+
+void CDisplayWindowGUI::createSubWindowsControlUI()
+{
+	constexpr int DEFAULT_WIDTH = 120, DEFAULT_HEIGHT = 40;
+	constexpr int HIDEN_WIDTH = 35, HIDEN_HEIGHT = 35;
+
+	if (m_subWindows.ui) return;
+
+	auto& w = m_subWindows.ui;
+	w = new nanogui::Window(this, "");
+
+	w->setLayout(new nanogui::BoxLayout(
+		nanogui::Orientation::Vertical, nanogui::Alignment::Fill, 1, 1));
+
+	w->setSize({DEFAULT_WIDTH, DEFAULT_HEIGHT});
+	w->setPosition({0, 0});
+	nanogui::Theme* modTheme = new nanogui::Theme(screen()->nvgContext());
+	modTheme->mWindowHeaderHeight = 1;
+	w->setTheme(modTheme);
+
+	auto stackPanels = w->add<nanogui::StackedWidget>();
+
+	auto page1 = new nanogui::Widget(nullptr);
+	auto page2 = new nanogui::Widget(nullptr);
+
+	stackPanels->addChild(0, page1);
+	stackPanels->addChild(1, page2);
+	stackPanels->setSelectedIndex(0);
+
+	page1->setLayout(new nanogui::BoxLayout(
+		nanogui::Orientation::Vertical, nanogui::Alignment::Fill, 1, 1));
+	page2->setLayout(new nanogui::BoxLayout(
+		nanogui::Orientation::Vertical, nanogui::Alignment::Fill, 1, 1));
+
+	auto btnUnhide =
+		page2->add<nanogui::Button>("", ENTYPO_ICON_CHEVRON_WITH_CIRCLE_RIGHT);
+	btnUnhide->setCallback([=]() {
+		stackPanels->setSelectedIndex(0);
+		w->setFixedSize({0, 0});
+		this->performLayout();
+	});
+	btnUnhide->setTooltip("Unhide this panel");
+
+	m_subWindows.uiCombo = page1->add<nanogui::ComboBox>();
+
+	m_subWindows.uiCombo->setCallback([this](int index) {
+		m_subWindows.restore(index);
+		m_subWindows.setFocused(index);
+	});
+
+	auto pn = page1->add<nanogui::Widget>();
+	pn->setLayout(new nanogui::BoxLayout(
+		nanogui::Orientation::Horizontal, nanogui::Alignment::Fill, 4, 4));
+
+	auto btnMin = pn->add<nanogui::Button>("", ENTYPO_ICON_ALIGN_BOTTOM);
+	btnMin->setCallback([this]() {
+		m_subWindows.minimize(m_subWindows.uiCombo->selectedIndex());
+	});
+	btnMin->setTooltip("Minimize");
+
+	auto btnMax = pn->add<nanogui::Button>("", ENTYPO_ICON_BROWSER);
+	btnMax->setCallback([this]() {
+		m_subWindows.restore(m_subWindows.uiCombo->selectedIndex());
+	});
+	btnMax->setTooltip("Restore selected");
+
+	auto btnMinAll = pn->add<nanogui::Button>("All", ENTYPO_ICON_ALIGN_BOTTOM);
+	btnMinAll->setCallback([this]() {
+		for (int i = 0; i < static_cast<int>(m_subWindows.windows.size()); i++)
+			m_subWindows.minimize(i);
+	});
+	btnMinAll->setTooltip("Minimize all");
+
+	auto btnMaxAll = pn->add<nanogui::Button>("All", ENTYPO_ICON_BROWSER);
+	btnMaxAll->setCallback([this]() {
+		for (int i = 0; i < static_cast<int>(m_subWindows.windows.size()); i++)
+			m_subWindows.restore(i);
+	});
+	btnMaxAll->setTooltip("Restore all");
+
+	auto btnHide =
+		pn->add<nanogui::Button>("", ENTYPO_ICON_CHEVRON_WITH_CIRCLE_LEFT);
+	btnHide->setCallback([=]() {
+		stackPanels->setSelectedIndex(1);
+		w->setFixedSize({HIDEN_WIDTH, HIDEN_HEIGHT});
+		this->performLayout();
+	});
+	btnHide->setTooltip("Hide this panel");
+}
+
+void CDisplayWindowGUI::SubWindows::minimize(int index)
+{
+	if (index < 0 || index > static_cast<int>(windows.size())) return;
+	auto w = windows.at(index);
+
+	w->setVisible(false);
+	// w->setFixedSize({1, 1});
+	parent.performLayout();
+}
+
+void CDisplayWindowGUI::SubWindows::restore(int index)
+{
+	if (index < 0 || index > static_cast<int>(windows.size())) return;
+	auto w = windows.at(index);
+
+	w->setVisible(true);
+	parent.performLayout();
+}
+
+void CDisplayWindowGUI::SubWindows::setFocused(int index)
+{
+	const int n = static_cast<int>(windows.size());
+	if (index >= 0 && index < n) windows.at(index)->requestFocus();
+}
+
 #endif  // MRPT_HAS_NANOGUI